LLM Frames Design/frame_engine_v1.1.2/code/pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "backend"
version = "1.0.0"
description = "Rodin Frame Engine Prototype"
license = "Proprietary"
authors = [
    "Charles-Edouard Bardyn <cbardyn@a-i.swiss>",
]
readme = "README.md"
packages = [
    { include = "backend", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11.8,<3.12"
PyYAML = "^6.0.1"
langchain = "^0.2.1"
langchain-google-genai = "^1.0.4"
langgraph = "^0.0.62"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"

# Optional LLM providers (install with: poetry install --extras "openai anthropic")
langchain-openai = { version = "^0.1.8", optional = true }
langchain-anthropic = { version = "^0.1.15", optional = true }

[tool.poetry.extras]
openai = ["langchain-openai"]
anthropic = ["langchain-anthropic"]
all-providers = ["langchain-openai", "langchain-anthropic"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
